braille_chewing_ime.py: Support backspace
When the backspace is pressed, the internal braille state is not reset.
The backspace is sent to the append_brl and the response is applied on the
underlying chewing input method. If the backspace is rejected due to an
empty braille buffer, it will be forwarded to the chewing input method
directly.

[!] Note that the backspace keydown message must be sent so that chewing
accepts it.