Make the dot pattern of ？ consistent with NVDA brl display
Reported by "嘯傲俠羽" <crazy@mail.batol.net>

The dot pattern of ？ (fullshape question mark) is 135-0 in the traditional
Chinese braille system, which is the same as the dot pattern of ） (fullshape
right parenthesis). This sometimes confuses users. Hence, ？ is defined as
1456-0 in the braille table of locale zh_TW in liblouis, a library serving
as one part of many screen reader for conversion between strings and
braille patterns.

In implementation of braille-pattern input, 視窗導盲鼠 (G-Mouse) and
高點 use different patterns. 視窗導盲鼠 uses 1345-0, which might be
learned by users when NVDA was not so popular. 高點 uses 1456-0, which
is consistent with NVDA's braille display.

Regarding that 視窗導盲鼠 is incompatible with Braille Chewing, it is
reasonable to change the braille pattern of ？ to decrease users' confusion.